Feat/user features (#3)
* Refactor exception handling; introduce AppException and specific exceptions for better error management

* Implement user management features; add user creation, profile retrieval, and update methods in UserFacade and UserService, along with corresponding request and response classes

* Feat: implement user registration and logout functionality; add Redis support for token management

* Feat: enhance category and product management; refactor response models, integrate MapStruct for mapping, and improve user profile update functionality

* Feat: refactor ProductController to use a constant for API tags and update SecurityConfig to expose public endpoints

* Enhance security configuration and logging; add role-based authorization, improve error handling

* Implement Cloudinary integration for image management; add services for uploading, retrieving, and deleting images

* Update ProductRepository to include category in entity graph for fixing N+1 problem

* Write unit tests for service classes

* Add unit tests for User, Product, and Category controllers; enhance security with role-based access

* Enhance security and validation: update error handling in application properties, refine authorization, add validation for user registration and login, and implement unauthorized access tests for product, category and user operations.

* Refactor JWT authentication handling: improve token validation response and enhance CORS configuration

* Enhance product specifications and error handling: improve name filtering, add active and featured specifications, and refine error responses in exception handling

* Enhance product details retrieval and response structure: add security requirements, implement role-based response for product details, and introduce new fields for product management

* Enhance product update functionality: allow image uploads during product updates, refine product response mapping, and improve category validation logic

* Implement mail service and password change functionality: add MailService interface and implementation, enhance UserService with password update method, and introduce OTP verification for secure password changes.

* Implement user activation functionality: add activateUser method in UserService, UserFacade, and UserServiceImpl, and create corresponding endpoint in UserController; fix test methods

+++ src/main/java/com/dungpham/asm1/repository/ProductRatingRepository.java
@@ -5,12 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface ProductRatingRepository extends JpaRepository<ProductRating, Long> {
     boolean existsByUserIdAndProductId(Long userId, Long productId);
 
     @Query("SELECT AVG(r.rating) FROM ProductRating r WHERE r.product.id = :productId")
-    Double findAverageRatingByProductId(@Param("productId") Long productId);
+    BigDecimal findAverageRatingByProductId(@Param("productId") Long productId);
 
 }
