+++ src/main/java/com/dungpham/asm1/service/ProductRatingService.java
@@ -3,7 +3,9 @@ package com.dungpham.asm1.service;
 import com.dungpham.asm1.entity.Product;
 import com.dungpham.asm1.entity.User;
 
+import java.math.BigDecimal;
+
 public interface ProductRatingService {
     void rateProduct(Product product, User user, int rating);
-    double getAverageRatingOfProduct(Product product);
+    BigDecimal getAverageRatingOfProduct(Product product);
 }
